Adding explicit dolphin/ prefix & fix characters

+++ src/JSystem/JFramework/JFWDisplay.cpp
@@ -9,6 +9,7 @@
 #include "dolphin/gx.h"
 #include <dolphin/vi.h>
 #include "global.h"
+#include "stdint.h"
 
 void JFWDisplay::ctor_subroutine(bool enableAlpha) {
     mEnableAlpha = enableAlpha;
@@ -362,8 +363,8 @@ static void waitForTick(u32 p1, u16 p2) {
             if (!OSReceiveMessage(JUTVideo::getManager()->getMessageQueue(), &msg, OS_MESSAGE_BLOCK)) {
                 msg = 0;
             }
-        } while (((int)msg - (int)nextCount) < 0);
-        nextCount = (int)msg + uVar1;
+        } while (((intptr_t)msg - (intptr_t)nextCount) < 0);
+        nextCount = (intptr_t)msg + uVar1;
     }
 }
